fileprocessor/requirements.txt
dnspython==2.1.0
pymongo==3.11.4
python-dotenv==0.18.0
